+++ tests/spec/core_functions/meta/load_css/error.rs
@@ -166,7 +166,6 @@ mod member {
         );
     }
     #[test]
-    #[ignore] // wrong error
     fn namespace() {
         let runner = runner().with_cwd("namespace");
         assert_eq!(